508965 - Rounds up obligation calculation result and changes tonne property type from float to integer

+++ src/EPR.PRN.Backend.Data/DataModels/ObligationCalculation.cs
@@ -26,5 +26,5 @@ public class ObligationCalculation
     public DateTime CalculatedOn { get; set; }
 
     [Required]
-    public double Tonnage { get; set; }
+    public int Tonnage { get; set; }
 }
